Remove getOwningTaskName method from InputGate
In order to make abstract InputGate simple for extending new implementations in shuffle service architecture, we could remove unnecessary methods from it.
InputGate#getOwningTaskName is only used for debugging log in BarrierBuffer and StreamInputProcessor. This task name could be got from Environment#getTaskInfo
and then be passed into the constructor of BarrierBuffer/StreamInputProcessor for use.