Refactor IOManager to extract internal FileChannelManager
IOManager mainly has two roles. One is for managing file channels based on config temp dirs, and the other is for abstracting ways to read/writer files.
We could define an internal FileChannelManager class for handing the file channels which could be reused for shuffle environment future. To do so the
shuffle environment do not need to rely on the whole IOManager.