Reorder plugins
If the scala 2.12 profile is active then the build-helper-maven-plugin is defined in the parent pom. This has odd effects about the order of plugins in the effective pom, resulting in the javacc plugin being executed before the fmpp plugin.
It looks like all modules after the build-helper-maven-plugin were moved to the front, so we're moving the plugin to the bottom to ensure the javacc/fmpp plugins aren't messed with.