Remove abstract getPageSize method from InputGate
Currently InputGate#getPageSize is only used for constructing BarrierBuffer. In order to make abstract InputGate simple and clean, we should remove unnecessary abstract methods from it.
Considering the page size could be parsed directly from configuration which could also visible while constructing BarrierBuffer, so it is reasonable to do so.