Fix Max_Retract and Min_Retract may produce incorrect result
The reason is we didn't set null to acc.max/min which may have an old value when we need to get a new max/min from an empty MapView.
And the old value will be output to downstream instead of a null value. This influences the final result.

This causes many unstable cases, including:
- AggregateITCase.testNestedGroupByAgg
- SplitAggregateITCase.testMinMaxWithRetraction

This closes #9120