temporarily disable liveserve
./build_docs.sh -i previously did not only enable incremental documentation
building while serving the docs, it also enabled a 'liveserve' mode that
automatically reloaded pages in the browser when they changed. This is based
on the 'hawkins' module which is not (yet) compatible with jekyll 4.0 which we
need to (significantly) improve build times.

This disables the liveserve mode and remove the hawkins module until a new
version is available.