Remove taskName parameter from StreamTask#createRecordWriter
We already pass the Environment as parameter in StreamTask#createRecordWriter, and the task name could be got directly form
Environment. It is better to pass individual dependent parameters instead of heavy-weight Environment, but it would bring
more refactoring works. So we keep the previous way and only remove task name at the moment.