Decouple OperatorChain from StreamStatusMaintainer
There are two motivations for this refactoring:
1. It is the precondition for the following work of decoupling the dependency between two inputs status in ForwardingValveOutputHandler.
2. From the aspect of design rule, the current OperatorChain takes many unrelated roles like StreamStatusMaintainer to make it unmaintainable.
The root reason for this case is from the cycle dependency between RecordWriterOutput (created by OperatorChain) and  StreamStatusMaintainer.

The solution is to refactor the creation of StreamStatusMaintainer and RecordWriterOutput in StreamTask level, and then break the implementation
cycle dependency between them. The array of RecordWriters which has close relationship with RecordWriterOutputs is created in StreamTask, so it
is reasonable to create them together. The created StreamStatusMaintainer in StreamTask can be directly referenced by subclasses like
OneInputStreamTask/TwoInputStreamTask.