Pass the counter of numRecordsIn into the constructors of StreamOne/TwoInputProcessor
Currently the counter of numRecordsIn is setup while processing input in processor. In order to integrate the processing logic based on
StreamTaskInput#emitNext(Output) later, we need to pass the counter into output functions then. So this refactoring is the precondition
of following works, and it could get additional benefits. One is that we could make the counter as final field in StreamInputProcessor.
Another is that we could reuse the counter setup logic for both StreamOne/TwoInputProcessors. There should be no side effects if we make
the counter setup a bit earlier than the previous way.