Migrate isFinished method from AvailabilityListener to AsyncDataInput
AvailabilityListener is used both in AsyncDataInput and StreamTaskInput. We already introduced InputStatus for PushBasedAsyncDataInput#emitNext,
and then InputStatus#END_OF_INPUT has the same semantic with AvailabilityListener#isFinished. But for the case of AsyncDataInput which is mainly
used by InputGate layer, the isFinished() method is still needed at the moment. So we migrate this method from AvailabilityListener to
AsyncDataInput, and refactor the StreamInputProcessor implementations by using InputStatus to judge the finished state.